Rename storage to session
