+++ src/Service/Auth/Authenticator/OAuth2Authenticator.php
@@ -7,7 +7,7 @@ use Ridibooks\Cms\Auth\LoginService;
 use Ridibooks\Cms\Service\Auth\Exception\NoCredentialException;
 use Ridibooks\Cms\Service\Auth\OAuth2\Client\OAuth2ClientInterface;
 use Ridibooks\Cms\Service\Auth\OAuth2\Exception\InvalidStateException;
-use Ridibooks\Cms\Service\Auth\Storage\SessionStorageInterface;
+use Ridibooks\Cms\Service\Auth\Session\SessionStorageInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 class OAuth2Authenticator extends BaseAuthenticator
